refactor: sessions code refactored out separately
